Changed longs to ints and naming

+++ CSGO/CSGO/Source/Program/global.h
@@ -100,7 +100,7 @@ struct vector_t
 	vector_t( coordinate_t, coordinate_t );
 };
 
-void Wait( unsigned long );
+void Wait( unsigned int );
 moment GetMoment( );
 
 template< typename xDatatype, typename xDatatype2 > void Limit( xDatatype &xStatus, xDatatype2 xMinimum, xDatatype2 xMaximum )
